+++ python/interpret_community/widget/js/.babelrc
@@ -1,0 +1,4 @@
+{
+    "presets": ["@babel/preset-env", "@babel/preset-react"],
+    "plugins": ["@babel/plugin-proposal-class-properties"]
+  }
