--- python/interpret_community/widget/js/.gitignore
@@ -1,14 +1,0 @@
-*.egg-info/
-.ipynb_checkpoints/
-
-build/
-*.py[cod]
-node_modules/
-/lib/
-
-# js dependencies
-/node_modules/
-
-
-# OS X
-.DS_Store
