--- python/interpret_community/widget/js/.npmrc
@@ -1,3 +1,0 @@
-@vienna:registry=https://msdata.pkgs.visualstudio.com/_packaging/vienna-shared-ux/npm/registry/
-always-auth=true
-
