--- python/interpret_community/widget/js/LICENSE
@@ -1,15 +1,0 @@
-This Preview is made available to you on the condition that you agree to the
-[Supplemental Terms of Use for Microsoft Azure Previews][1], which supplement
-[your agreement][2] governing your use of Azure.
-If you do not have an existing agreement governing your use of Azure, you agree that 
-your agreement governing use of Azure is the [Microsoft Online Subscription Agreement][3]
-(which incorporates the [Online Services Terms][4]).
-By using the Preview you agree to these terms. This Preview may collect data
-that is transmitted to Microsoft. Please see the [Microsoft Privacy Statement][5]
-to learn more about how Microsoft processes personal data.
-
-[1]: https://azure.microsoft.com/en-us/support/legal/preview-supplemental-terms/
-[2]: https://azure.microsoft.com/en-us/support/legal/
-[3]: https://azure.microsoft.com/en-us/support/legal/subscription-agreement/
-[4]: http://www.microsoftvolumelicensing.com/DocumentSearch.aspx?Mode=3&DocumentTypeId=46
-[5]: http://go.microsoft.com/fwlink/?LinkId=248681
