--- python/interpret_community/widget/js/README.md
@@ -1,11 +1,0 @@
-Machine Learning explanation widget for Jupyter widget
-
-Package Install
----------------
-
-**Prerequisites**
-- [node](http://nodejs.org/)
-
-```bash
-npm install --save interpret-ml-widget
-```
