+++ python/interpret_community/widget/js/package.json
@@ -1,60 +1,31 @@
 {
-  "name": "interpret-ml-widget",
-  "version": "0.1.7",
-  "description": "Machine Learning explanation widget for Jupyter notebooks",
-  "author": "Interpret ML",
-  "license": "See license file 'LICENSE'",
-  "main": "lib/index.js",
-  "types": "./lib/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/microsoft/interpret"
-  },
-  "keywords": [
-    "jupyter",
-    "widgets",
-    "ipython",
-    "ipywidgets",
-    "jupyterlab-extension"
-  ],
-  "files": [
-    "lib/**/*.js",
-    "dist/*.js"
-  ],
-  "scripts": {
-    "build": "npm run build:lib && npm run build:nbextension",
-    "build:labextension": "npm run clean:labextension && mkdirp interpret_ml_widget/labextension && cd interpret_ml_widget/labextension && npm pack ../..",
-    "build:lib": "tsc",
-    "build:nbextension": "webpack -p",
-    "build:all": "npm run build:lib && npm run build:labextension && npm run build:nbextension",
-    "clean": "npm run clean:lib && npm run clean:nbextension",
-    "clean:lib": "rimraf lib",
-    "clean:labextension": "rimraf interpret_ml_widget/labextension",
-    "clean:nbextension": "rimraf interpret_ml_widget/nbextension/static/index.js"
-  },
-  "devDependencies": {
-    "@phosphor/application": "^1.6.0",
-    "@phosphor/widgets": "^1.6.0",
-    "@types/node": "^10.11.6",
-    "@types/react-dom": "^16.8.4",
-    "mkdirp": "^0.5.1",
-    "rimraf": "^2.6.1",
-    "source-map-loader": "^0.2.4",
-    "ts-loader": "^5.2.1",
-    "typescript": "~3.1.2",
-    "webpack": "4.30.0",
-    "webpack-cli": "^3.3.1"
-  },
-  "dependencies": {
-    "@jupyter-widgets/base": "^1.1.10",
-    "css-loader": "^2.1.1",
-    "lodash": "^4.17.11",
-    "mlchartlib": "0.1.31",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "style-loader": "^0.23.1"
-  },
-  "jupyterlab": {
-    "extension": "lib/labplugin"
+    "name": "component-host",
+    "version": "1.0.0",
+    "description": "Host for component to be consumed as npm package",
+    "main": "main.js",
+    "scripts": {
+      "test": "echo \"Error: no test specified\" && exit 1",
+      "build": "webpack",
+      "start": "start http://localhost:7777/ & webpack-dev-server"
+    },
+    "author": "",
+    "license": "MIT",
+    "dependencies": {
+      "@babel/core": "7.4.3",
+      "@babel/plugin-proposal-class-properties": "^7.4.0",
+      "@babel/preset-env": "7.4.3",
+      "@babel/preset-react": "7.0.0",
+      "babel-loader": "8.0.5",
+      "babel-preset-es2015": "6.24.1",
+      "babel-preset-react": "6.24.1",
+      "css-loader": "2.1.1",
+      "mlchartlib": "0.1.31",
+      "react": "16.8.6",
+      "react-dom": "16.8.6",
+      "style-loader": "0.23.1",
+      "webpack": "4.29.6",
+      "webpack-cli": "3.3.0",
+      "webpack-dev-server": "3.2.1"
+    }
   }
-}
+
