--- python/interpret_community/widget/js/src/index.ts
@@ -1,3 +1,0 @@
-const data = require('../package.json');
-export const version = data.version;
-export * from './explanationDashboard';
