--- python/interpret_community/widget/js/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "esModuleInterop":true,
-    "lib": ["es2015", "dom"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmitOnError": true,
-    "noUnusedLocals": true,
-    "outDir": "lib",
-    "resolveJsonModule": true,
-    "rootDir": "src",
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "strict": true,
-    "target": "es2015",
-    "jsx": "react",
-    "noImplicitAny": false
-  },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.tsx",
-  ]
-}
