--- python/interpret_community/widget/njs/.babelrc
@@ -1,4 +1,0 @@
-{
-    "presets": ["@babel/preset-env", "@babel/preset-react"],
-    "plugins": ["@babel/plugin-proposal-class-properties"]
-  }
