--- python/interpret_community/widget/njs/main.js
@@ -1,33 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-import { ExplanationDashboard } from 'mlchartlib';
-
-
-let generatePrediction = (postData) => {
-  return fetch(data.predictionUrl, {method: "post", body: JSON.stringify(postData), headers: {
-    'Content-Type': 'application/json'
-  }}).then(resp => {
-    if (resp.error !== undefined) {
-      throw new Error(resp.error)
-    }
-    return resp.data
-  })
-}
-
-ReactDOM.render(<ExplanationDashboard
-    modelInformation={{modelClass: 'blackbox'}}
-    dataSummary={{featureNames: data.featureNames, classNames: data.classNames}}
-    testData={data.trainingData}
-    predictedY={data.predictedY}
-    probabilityY={data.probabilityY}
-    trueY={data.trueY}
-    precomputedExplanations={{
-      localFeatureImportance: data.localExplanations,
-      globalFeatureImportance: data.globalExplanation,
-      ebmGlobalExplanation: data.ebmData
-    }}
-    requestPredictions={data.predictionUrl !== undefined ? generatePrediction : undefined}
-    theme={"dark"}
-    key={new Date()}
-  />, document.getElementById(data.divId));
